fix(media): keep route exports Next-compatible
