+++ app/media/[...key]/route.ts
@@ -1,12 +1,10 @@
 import { getCloudflareContext } from "@opennextjs/cloudflare";
+import { resolvePublicMediaKey } from "@/lib/media/public-key";
 
 interface RouteContext {
   params: Promise<{ key?: string[] }>;
 }
 
-const PUBLIC_MEDIA_PREFIXES = new Set(["products", "categories", "blog", "pages"]);
-const SAFE_KEY_SEGMENT = /^[A-Za-z0-9][A-Za-z0-9._~-]*$/;
-const MAX_KEY_LENGTH = 1024;
 const REVALIDATING_CACHE = "public, max-age=0, must-revalidate";
 const INERT_MEDIA_CSP = "default-src 'none'; sandbox";
 
@@ -29,28 +27,6 @@ function simpleResponse(message: string, status: number, head = false): Response
       "X-Content-Type-Options": "nosniff",
     },
   });
-}
-
-/** Next has already decoded dynamic route params; never decode them again. */
-export function resolvePublicMediaKey(segments: readonly string[] | undefined): string | null {
-  if (!segments || segments.length < 2 || !PUBLIC_MEDIA_PREFIXES.has(segments[0])) return null;
-  if (segments.length > 12) return null;
-
-  for (const segment of segments) {
-    if (
-      segment.length > 255
-      || !SAFE_KEY_SEGMENT.test(segment)
-      || segment === "."
-      || segment === ".."
-      || segment.includes("%")
-      || segment.includes("\\")
-    ) {
-      return null;
-    }
-  }
-
-  const key = segments.join("/");
-  return key.length <= MAX_KEY_LENGTH ? key : null;
 }
 
 function verifiedContentType(object: R2Object): string | null {
